requirements.txt
paho-mqtt
dotenv
requests
